+++ api/v1/class/mode/mode.arbeit.inc.php
@@ -21,8 +21,15 @@ namespace Arbeitszeit {
         public static function compute_html_worktime_types()
         {
             $data = "";
+            $selected = "";
             foreach (Arbeitszeit::get_all_types() as $type => $value) {
-                $data .= "<option value=\"{$type}\">{$value}</option>";
+                $selected = "";
+
+                if((int)$type == Arbeitszeit::get_app_ini()["config"]["default_worktime_type"]) {
+                    $selected = " selected";
+                }
+                print_r($type);
+                $data .= "<option value=\"{$type}\"{$selected}>{$value}</option>";
             }
             return $data;
         }